Separation Reliability: Person Separation Reliability

Copied from eRm::SepRel(), v1.0-1: This function calculates the proportion of person variance that is not due to error. The concept of person separation reliability is very similar to reliability indices such as Cronbach's alpha.

SepRel_1.0.1(pobject)

Arguments

pobject

Object of class ppar (see ?eRm::person.parameter).

Value

SepRel returns a list object of class eRm_SepRel containing:

sep.rel

the person separation reliability,

SSD.PS

the squared standard deviation (i.e., total person variability),

MSE

the mean square measurement error (i.e., model error variance).

Details

See full documentation at https://www.rdocumentation.org/packages/eRm/versions/1.0-1/topics/Separation%20Reliability

References

Wright, B.D., and Stone, M.H. (1999). Measurement essentials. Wide Range Inc., Wilmington. (http://www.rasch.org/measess/me-all.pdf 28Mb).

Author

Original code by Adrian Brügger (Adrian.Bruegger@imu.unibe.ch), adapted by Marco J. Maier in package eRm v1.0-1