Example of WHO Model Disability Survey data for adults

An dummy data set of data from the WHO Model Disability Survey. All survey variables and demographic characteristics (except work status) are randomly generated. Responses to the survey questions and work status are from randomly selected rows of the 2014 Chilean implementation of the MDS, ENDIS II (2014).

df_adults

Format

A tibble with 2500 rows and 90 variables:

HHID

household number

strata

survey strata

PSU

PSU id

weight

survey weights

sex

sex, Male or Female

age

age in years

age_cat

age catgory, one of "18-24", "25-39", "40-64", or "64-100"

edu_cat

highest level of education attained

work_cat

whether respondent is currently working, binary

disability_score

metric scale of performance, from Rasch analysis

disability_cat

performance level, as definted by WHO cutoffs of disability_score

Functioning section: On a scale from 1 (none) to 5 (extreme), in the last 30 days, taking into account any help or support you receive, how much has of a problem has it been for you to...
F1

...stand up from sitting?

F2

...stand for long periods of time, for example 30 minutes?

F3

...leave the house?

F4

...walk short distances, such as a street block or 100 meters?

F5

...walk 10 blocks or a kilometer?

F6

...do vigorous activities, such as play footbal, lift heavy objects, ride a bike or run?

F7

...get where you want to go?

F8

...manipulate small objects or opening a container?

F9

...lift a 2L full bottle of water from your waist to eye level?

F10

...wash or dress yourself?

F11

...feed yourself?

F12

...use the bathroom?

F13

...cut your toenails?

F14

...take care of your health, for example exercise, eat well or take your medications?

F15

...see objects at a distance?

F16

...see objects at arms length?

F17

...hear a conversation in a quiet room?

F18

...hear a conversation in a loud room?

F19

...feel pain?

F20

...sleep?

F21

...feel tired or not have enough energy?

F22

...feel short of breath?

F23

...cough or wheeze?

F24

...feel sad, down or depressed?

F25

...feel worried, nervous or anxious?

F26

...get along with people close to you, including your family and friends?

F27

...get along with people you don't know?

F28

...make new friends or maintain your friendships?

F29

...have intimiate relationships?

F30

...manage stress?

F31

...cope with everything you have to do?

F32

...be understood in your usual language?

F33

...understand others in your usual language?

F34

...forget things?

F35

...remember the important things you have to do day-to-day?

F36

...find solutions to your day-to-day problems?

F37

...complete household tasks, like sweeping, cooking, arranging the house or taking our the trash?

F38

...manage the money you have?

F39

...do things for relaxation?

F40

...participate in community activities?

F41

...participate in local or national policitics or civil society?

F42

...take care of others?

F43

...get a job?

F44

...obtain a higher education?

F45

...use public transportation?

F46

...get things done in your job (if not currently working, NA)?

F47

...get things done at your school (if not currently studying, NA)?

Capacity section: On a scale from 1 (none) to 5 (extreme), in the last 30 days, withing taking into account any type of help or support, due to your health how much difficulty have you had with...
C2

...seeing, without contact lenses or glasses?

C3

...hearing, without hearing aids?

C4

...walking or climbing steps?

C5

...remembering or concentrating?

C6

...washing or dressing?

C7

...communicating in your usual language?

C8

...feeding yourself?

C9

...using the bathroom?

C10

...waking up and getting out of bed?

C11

...going out to the street?

C12

...doing shopping or going to the doctor?

C13

...manipulating small objects or opening a container?

C14

...sleeping?

C15

...breathing?

C16

...doing household tasks like sweeping, cooking, arranging the house or taking out the trash?

C17

...taking care of others?

C18

...participating in community activities?

C19

...feeling sad, down or depressed?

C20

...feeling worried, nervous or anxious?

C21

...getting along with people close to you, including your family and friends?

C22

...coping with everything you have to do?

C23

...feeling pain?

C24

...getting things done in your job (if not currently working, NA)?

C25

...getting things done at your school (if not currently studying, NA)?

Environmental factors section: On a scale from 1 (very easy) to 5 (very hard), to what extent...
EF1

...do health facilities you need regularly make it easy or hard for you to use them?

EF2

...do places where you socialize and engage in community activities make it easy or hard for you to do this?

EF3

...do the shops, banks and post office in your neighbourhood make it easy or hard for you to use them?

EF4

...do your regular places of worship make it easy or hard for you to worship?

EF5

...does the transportation you need or want to use make it easy or hard for you to use it?

EF6

...does your dwelling (including the toilet and all rooms) make it easy or hard for you to live there?

EF7

...do the temperature, terrain, and climate of the place you usually live make it easy or hard for you to live there?

EF8

...does the lighting in your surroundings make it easy or hard for you to live there?

EF9

...does the noise in your surroundings make it easy or hard for you to live there?

EF10

...do the crowds in your surroundings make it easy or hard for you to live there?

EF11

...does your workplace make it easy or hard for you to work or learn (if not currently working, NA)?

EF12

...does your educational institution make it easy or hard for you to work or learn (if not currently studying, NA)?