Example of WHO Model Disability Survey data for children

An dummy data set of data from the WHO Model Disability Survey for children. All survey variables and demographic characterstics (except age variables) are randomly generated. Responses to the survey questions and age variables are from randomly selected rows of the 2014 Chilean implementation of the MDS, ENDIS II (2014).

df_children

Format

A tibble with 2500 rows and 42 variables:

HHID

household number

strata

survey strata

PSU

PSU id

weight

survey weights

sex

sex, Male or Female

age

age in years

age_cat

age catgory, one of "Age2to4", "Age5to9", or "Age10to17"

Functioning section: On a scale from 1 (none) to 5 (extreme), in the last 30 days, taking into account any help or support the child receives, compared with children of the same age, how much has of a problem it been for the child to...
child1

...walk?

child2

...manipulat small objects or opening a container?

child3

...see things from a long distance?

child4

...hear?

child5

...feel pain?

child6

...not have enough energy?

child7

...have too much energy

child8

...feel short of breath?

child9

...feel sad, down or depressed?

child10

...feel worried, nervous or anxious?

child11

...(for children aged 5 to 17) wash or dress him- or herself?

child12

...(for children aged 2 to 4) bite or hit other children or adults?

child13

...(for children aged 5 to 17) control his or her own behavior?

child14

...(for children aged 5 to 17) get along with children of the same age?

child15

...(for children aged 2 to 4) understand what you say to him or her?

child16

(for children aged 2 to 4) How much of a problem has it been for you to understand what the child says?

child17

...(for children aged 5 to 17) understand other people?

child18

...(for children aged 5 to 17) be understood?

child19

...(for children aged 2 to 3) learn the names of household objects?

child20

...(for children aged 3 to 17) learn to do new things?

child21

...(for children aged 5 to 17) complete a task?

child22

...(for children aged 5 to 17) make changes to his or her routine?

child23

...(for children aged 5 to 17) do homework as requested at school?

child24

...(for children aged 2 to 5) play with toys or domestic objects?

child25

...(for children aged 2 to 12) play with other children?

child26

...(for children aged 13 to 17) do activities with other children?

child27

...(for children aged 5 to 17) participate in community activities?

Capacity section: On a scale from 1 (none) to 5 (extreme), in the last 30 days, without taking into account any help or support the child receives, how much has difficulty has the child had with...
child28

...seeing without glasses?

child29

...hearing without hearing aids?

child30

...walking?

child31

...understanding you or others?

child32

...learning?

child33

...controling his or her behavior?

child34

...completing a task?

child35

...getting along with other children?