Create indicators from data frame

Create indicators from data frame

helper_indicator(
  df,
  vars_indicators,
  mapvalues_from,
  mapvalues_to,
  make_factor = TRUE
)

Arguments

df

a data frame

vars_indicators

a character vector of the variables from df to create indicators for

mapvalues_from

vector to pass to plyr::mapvalues() argument from

mapvalues_to

vector to pass to plyr::mapvalues() argument to

make_factor

a logical indicating whether resulting indicators should be factors. Default is TRUE.

Value

a data frame with new columns that are the indicators from vars_indicators, with the same names pasted with "_ind".

Examples

helper_indicator(df = df_adults, 
vars_indicators = c("EF1", "EF2", "EF3"), 
mapvalues_from = 1:5, 
mapvalues_to = c(0,0,0,1,1))
#> # A tibble: 2,500 × 93
#>     HHID strata   PSU weight sex     age age_cat edu_cat work_…¹ disab…² disab…³
#>    <dbl> <chr>  <int>  <dbl> <fct> <int> <ord>   <ord>   <fct>     <dbl> <ord>  
#>  1     1 B         19   1000 Fema…    37 25-39   Elemen… Y          43.0 Modera…
#>  2     2 C         11    800 Male     39 25-39   None    Y          21.8 Mild   
#>  3     3 C         14   1150 Fema…    67 64-100  Elemen… Y          NA   NA     
#>  4     4 B         11    900 Male     57 40-64   None    Y          43.0 Modera…
#>  5     5 B          4    850 Fema…    20 18-24   None    N           0   No     
#>  6     6 C          7   1200 Male     67 64-100  Elemen… Y          55.1 Severe 
#>  7     7 A          1    800 Male     37 25-39   Elemen… Y          48.2 Modera…
#>  8     8 B         20   1000 Male     44 40-64   None    N          NA   NA     
#>  9     9 B         18   1000 Male     89 64-100  Second… N          54.9 Severe 
#> 10    10 C         20   1100 Male     76 64-100  Elemen… Y          57.2 Severe 
#> # … with 2,490 more rows, 82 more variables: F1 <dbl>, F2 <dbl>, F3 <dbl>,
#> #   F4 <dbl>, F5 <dbl>, F6 <dbl>, F7 <dbl>, F8 <dbl>, F9 <dbl>, F10 <dbl>,
#> #   F11 <dbl>, F12 <dbl>, F13 <dbl>, F14 <dbl>, F15 <dbl>, F16 <dbl>,
#> #   F17 <dbl>, F18 <dbl>, F19 <dbl>, F20 <dbl>, F21 <dbl>, F22 <dbl>,
#> #   F23 <dbl>, F24 <dbl>, F25 <dbl>, F26 <dbl>, F27 <dbl>, F28 <dbl>,
#> #   F29 <dbl>, F30 <dbl>, F31 <dbl>, F32 <dbl>, F33 <dbl>, F34 <dbl>,
#> #   F35 <dbl>, F36 <dbl>, F37 <dbl>, F38 <dbl>, F39 <dbl>, F40 <dbl>, …