Run the Rasch Model and print diagnostic results

Run the Rasch Model and print diagnostic results

rasch_model(
  df,
  vars_metric,
  vars_id,
  print_results = FALSE,
  path_output = NULL,
  LIDcutoff = 0.2
)

Arguments

df

a data frame of individual survey data, where each row is an individual

vars_metric

a character vector of items to use in the Rasch Analysis

vars_id

a string with column name uniquely identifying individuals

print_results

a logical value indicating whether or not to print various files displaying results from the Rasch Model. Default is FALSE, to not print the files.

path_output

a string with the path to the output folder. Default is NULL.

LIDcutoff

either a numeric value between 0 and 1 indicating the cut-off for significant local item dependence, or the string "christensen" to use the cut-off suggested by Christensen et al. 2017 (see reference). If "christensen" cut-off fails, defaults to 0.2.

Value

a list with results from the Rasch Model:

model

the results from the Rasch Model

df_score

a tibble with the items used in the analysis and the person abilities

thresholds

the item thresholds (i.e., crossings)

person_parameters

person abilities

PSI

the person-separation index

item_fit

infit and outfit statistics per item

residuals_PCM

the standardized person residuals

LID

matrix with the item residual correlations

targeting

a matrix with information on the targeting of the model

fit_results

a string with results of the item fit

LID_results

a string with results of the local item dependency

disordered_results

a string listing items with disordered thresholds

Details

The Rasch Model is calculated using the function eRm::PCM().

References

Christensen, K. B., Makransky, G., & Horton, M. (2017). Critical Values for Yen’s Q 3 : Identification of Local Dependence in the Rasch Model Using Residual Correlations. Applied Psychological Measurement, 41(3), 178-194. doi:10.1177/0146621616677520

See also