Recode survey items for use in Rasch Analysis

Recode survey items for use in Rasch Analysis

rasch_recode(df, vars_metric, recode_strategy, max_values)

Arguments

df

a data frame of individual survey data, where each row is an individual

vars_metric

a character vector of items to use in the Rasch Analysis

recode_strategy

a named list giving the strategy to take for recoding variables, passed to rasch_recode(). One element of the list per recode strategy. Each element of the list is a numeric vector giving the new values to map the variables to. The names of the list are the groups of column names to use for each recoding strategy, separated only by ",". Default is NULL, to not recode items.

max_values

a tibble with two columns, var equivalent to vars_metric and max_val with their corresponding maximum possible values

Value

a named list with:

df

new df after recoding the desired variables

max_values

new max_values after recoding the desired variables

See also