Compute basic statistics of the number of members per group per household

Compute basic statistics of the number of members per group per household

table_basicstats(df, hh_id, group_by_var)

Arguments

df

a data frame of household data where the rows represent members of the households in the sample

hh_id

string (length 1) indicating the name of the variable in df uniquely identifying households

group_by_var

string (length 1) to pass to group_by_at() with name of variable in df to group results by

Value

A tibble with rows for each level of group_by_var and "Total" and columns for the Mean (SD), Median and Range of the number of people in each group per household.

Note

Includes a call to tidyr::complete(), which causes the function to be a bit slow.

See also

Examples

#create dummy table of household data, where each row represents one member
df_hh <- data.frame(HHID = sample(
  x = 1:300,
  size = 1000,
  replace = TRUE
),
age_cat = ordered(sample(
  x = c("18-24", "25-39", "40-64", "64-100"),
  size = 1000,
  replace = TRUE
)))
                
table_basicstats(df_hh, "HHID", "age_cat")
#> # A tibble: 5 × 4
#>   age_cat mean_sd   median range 
#>   <chr>   <chr>      <dbl> <chr> 
#> 1 18-24   0.9 (1)        1 0 - 4 
#> 2 25-39   0.8 (1)        1 0 - 5 
#> 3 40-64   0.9 (1)        1 0 - 5 
#> 4 64-100  0.9 (0.9)      1 0 - 4 
#> 5 Total   3.4 (1.8)      3 1 - 10