Compute unweighted percent and N for multiple variables, disaggregated

Compute unweighted percent and N for multiple variables, disaggregated

table_unweightedpctn(
  df,
  vars_demo,
  group_by_var = NULL,
  spread_by_group_by_var = FALSE,
  group_by_var_sums_to_100 = FALSE,
  add_totals = FALSE
)

Arguments

df

a data frame of individual survey data, where each row is an individual

vars_demo

a character vector of names of variables to calculate percent and N for

group_by_var

a string (length 1) with the name of the variable from df to disaggregate by

spread_by_group_by_var

logical determining whether to pass group_by_var to tidyr::pivot_wider() to give a wide-format tab. Default is FALSE.

group_by_var_sums_to_100

logical determining whether percentages sum to 100 along the margin of group_by_var, if applicable. Default is FALSE.

add_totals

logical determining whether to create total rows or columns (as appropriate) that demonstrate the margin that sums to 100. Default is FALSE.

Value

A tibble with percent and N for each level of each variable in vars_demo

See also

Examples

table_unweightedpctn(df_adults, vars_demo = c("sex", "age_cat", "work_cat", "edu_cat"))
#> # A tibble: 12 × 4
#>    item     demo         pct     n
#>    <chr>    <ord>      <dbl> <dbl>
#>  1 sex      Female      50.6  1265
#>  2 sex      Male        49.4  1235
#>  3 age_cat  18-24        7.6   191
#>  4 age_cat  25-39       18     449
#>  5 age_cat  40-64       31.7   792
#>  6 age_cat  64-100      42.7  1068
#>  7 work_cat N           41.8  1045
#>  8 work_cat Y           58.2  1455
#>  9 edu_cat  None        20.2   504
#> 10 edu_cat  Elementary  39.8   996
#> 11 edu_cat  Secondary   25.2   629
#> 12 edu_cat  University  14.8   371
table_unweightedpctn(df_adults, vars_demo = c("sex", "age_cat", "work_cat", "edu_cat"), 
group_by_var = "disability_cat")
#> # A tibble: 48 × 5
#>    item    demo   disability_cat   pct     n
#>    <chr>   <ord>  <ord>          <dbl> <dbl>
#>  1 sex     Female No              52.5   287
#>  2 sex     Male   No              47.5   260
#>  3 sex     Female Mild            53.5   250
#>  4 sex     Male   Mild            46.5   217
#>  5 sex     Female Moderate        49.9   376
#>  6 sex     Male   Moderate        50.1   377
#>  7 sex     Female Severe          49.1   209
#>  8 sex     Male   Severe          50.9   217
#>  9 age_cat 18-24  No               6.9    38
#> 10 age_cat 25-39  No              19.7   108
#> # ℹ 38 more rows
table_unweightedpctn(df_adults, vars_demo = c("sex", "age_cat", "work_cat", "edu_cat"), 
group_by_var = "disability_cat", spread_by_group_by_var = TRUE)
#> # A tibble: 12 × 10
#>    item    demo  No_pct  No_n Mild_pct Mild_n Moderate_pct Moderate_n Severe_pct
#>    <chr>   <ord>  <dbl> <dbl>    <dbl>  <dbl>        <dbl>      <dbl>      <dbl>
#>  1 sex     Fema…   52.5   287     53.5    250         49.9        376       49.1
#>  2 sex     Male    47.5   260     46.5    217         50.1        377       50.9
#>  3 age_cat 18-24    6.9    38      7.3     34          8.8         66        7.7
#>  4 age_cat 25-39   19.7   108     16.1     75         19.7        148       16.4
#>  5 age_cat 40-64   32.4   177     33.2    155         29.3        221       32.6
#>  6 age_cat 64-1…   41     224     43.5    203         42.2        318       43.2
#>  7 work_c… N       41.7   228     40.5    189         42          316       42.3
#>  8 work_c… Y       58.3   319     59.5    278         58          437       57.7
#>  9 edu_cat None    21.8   119     16.7     78         21.4        161       19  
#> 10 edu_cat Elem…   37.8   207     38.8    181         39.7        299       41.3
#> 11 edu_cat Seco…   25     137     28.7    134         23.2        175       25.1
#> 12 edu_cat Univ…   15.4    84     15.8     74         15.7        118       14.6
#> # ℹ 1 more variable: Severe_n <dbl>