Calculate table of percentages or N of response distribution for survey items, survey weighted, disaggregated

Calculate table of percentages or N of response distribution for survey items, survey weighted, disaggregated

table_weightedpct(
  df,
  vars_ids,
  vars_strata,
  vars_weights,
  formula_vars,
  ...,
  formula_vars_levels = 0:1,
  by_vars = NULL,
  pct = TRUE,
  willfilter = NULL,
  add_totals = FALSE,
  spread_key = NULL,
  spread_value = "prop",
  arrange_vars = NULL,
  include_SE = FALSE
)

Arguments

df

a data frame of individual survey data, where each row is an individual

vars_ids

a character vector of cluster ids, passed to srvyr::as_survey_design()

vars_strata

a character vector of strata ids, passed to srvyr::as_survey_design()

vars_weights

a character vector of survey weight ids, passed to srvyr::as_survey_design()

formula_vars

a character vector of variables to calculate the percentages of each level for

...

captures expressions to pass to dplyr::filter() or dplyr::transmute(), depending on the value of argument willfilter. See Details.

formula_vars_levels

a vector of the levels of the the formula_vars

by_vars

a character vector of variables to disaggregate results by. Default is NULL for no disaggregation. The columns listed must not include NAs.

pct

a logical variable indicating whether or not to calculate weighted percentages. Default is TRUE for weighted percentages. Set to FALSE for weighted N.

willfilter

a logical variable that tells the function whether or not to filter or transmute the data. Leave as default NULL to not filter or transmute. Set as TRUE to filter and FALSE to transmute. See Details.

add_totals

logical determining whether to create total rows or columns (as appropriate) that demonstrate the margin that sums to 100. Default is FALSE.

spread_key

a string with variable name to pass to names_from argument of tidyr::pivot_wider(). Default is NULL.

spread_value

a string with variable name to pass to values_from argument of tidyr::pivot_wider(). Default is "prop" (the column of percentages created within the function)

arrange_vars

a character vector with variables to pass to dplyr::arrange(). Default is NULL.

include_SE

a logical variable indicating whether to include the standard errors in the table. Default is FALSE. Currently does not work when adding totals, spreading or transmuting.

Value

a tibble of weighted response percentages or N's

Details

If willfilter is NULL, the table is not filtered or transmuted. If willfilter is TRUE, the table is filtered before it is spread or arranged. If willfilter is FALSE, the table is transmuted after the spread and/or arrange. "..." captures the non-standard evaluation expressions (NSE) to pass to dplyr::filter or dplyr::transmute().

The function performs the following actions with the table after results are calculated in the following order (if applicable): filter, add totals, spread, arrange, transmute

See also

Examples

table_weightedpct(df_adults, 
    vars_ids = c("HHID", "PSU"),
    vars_strata = "strata",
    vars_weights = "weight",
    formula_vars = paste0("EF",1:10),
    formula_vars_levels = 1:5,
    by_vars = "sex")
#> Warning: You may have issues with lonely PSUs if you have not set: options(survey.lonely.psu = "adjust")
#> # A tibble: 100 × 4
#> # Groups:   sex, item [20]
#>    sex    item  resp   prop
#>    <fct>  <ord> <fct> <dbl>
#>  1 Female EF1   1     44.6 
#>  2 Female EF1   2     15.3 
#>  3 Female EF1   3     15.4 
#>  4 Female EF1   4     10.5 
#>  5 Female EF1   5     14.1 
#>  6 Female EF10  1     47.2 
#>  7 Female EF10  2     10.7 
#>  8 Female EF10  3     20.6 
#>  9 Female EF10  4     11.8 
#> 10 Female EF10  5      9.81
#> # … with 90 more rows