whomds: A package for calculating results from WHO Model Disability Survey

The whomds package provides three categories of important functions: table functions, figure functions, and Rasch Analysis functions

Table functions

The table functions output different fit for purpose tables for reporting results from the WHO Model Disability Survey (MDS). They begin with table_*()

Figure functions

The figures functions output different figures for reporting results from the WHO Model Disability Survey (MDS). They begin with fig_*()

Rasch Analysis functions

These functions are used to complete an iteration of Rasch Analysis for WHO Model Disability Survey (MDS). They begin with rasch_*

References

WHO Model Disability Survey: https://www.who.int/health-topics/disability

See also

Author

Maintainer: Lindsay Lee lindsayevanslee@gmail.com

Other contributors:

Carolina Fellinghauer carolina.fellinghauer@data-fittery.ch [contributor]

World Health Organization disability@who.int [copyright holder]